Refactor common AxisItem code from owfreeviz and owlinearprojection
Also fix compatibility with pyqtgraph 0.10 (`AttributeError: 'TextItem'
object has no attribute 'updateText'`)
